Move ShutdownHandler to core
Needs to be accessible from other modules which don't directly depend on the rest code.
