+++ core/src/main/java/org/apache/brooklyn/core/mgmt/ShutdownHandler.java
@@ -16,30 +16,8 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.brooklyn.rest.util;
+package org.apache.brooklyn.core.mgmt;
 
-import javax.annotation.Nullable;
-import javax.ws.rs.ext.ContextResolver;
-import javax.ws.rs.ext.Provider;
-
-import org.apache.brooklyn.core.mgmt.ShutdownHandler;
-
-
-@Provider
-public class ShutdownHandlerProvider implements ContextResolver<ShutdownHandler> {
-    private ShutdownHandler shutdownHandler;
-
-    public ShutdownHandlerProvider(@Nullable ShutdownHandler instance) {
-        this.shutdownHandler = instance;
-    }
-
-    @Override
-    public ShutdownHandler getContext(Class<?> type) {
-        if (type == ShutdownHandler.class) {
-            return shutdownHandler;
-        } else {
-            return null;
-        }
-    }
-
+public interface ShutdownHandler {
+    void onShutdownRequest();
 }
