+++ core/src/main/resources/org/apache/brooklyn/core/mgmt/persist/deserializingClassRenames.properties
@@ -1257,7 +1257,7 @@ brooklyn.rest.util.DefaultExceptionMapper                                       
 brooklyn.rest.util.EntityLocationUtils                                           : org.apache.brooklyn.rest.util.EntityLocationUtils
 brooklyn.rest.util.FormMapProvider                                               : org.apache.brooklyn.rest.util.FormMapProvider
 brooklyn.rest.util.ManagementContextProvider                                     : org.apache.brooklyn.rest.util.ManagementContextProvider
-brooklyn.rest.util.ShutdownHandler                                               : org.apache.brooklyn.rest.util.ShutdownHandler
+brooklyn.rest.util.ShutdownHandler                                               : org.apache.brooklyn.core.mgmt.ShutdownHandler
 brooklyn.rest.util.ShutdownHandlerProvider                                       : org.apache.brooklyn.rest.util.ShutdownHandlerProvider
 brooklyn.rest.util.json.BidiSerialization                                        : org.apache.brooklyn.rest.util.json.BidiSerialization
 brooklyn.rest.util.json.BrooklynJacksonJsonProvider                              : org.apache.brooklyn.rest.util.json.BrooklynJacksonJsonProvider
