+++ rest/rest-server/src/test/java/org/apache/brooklyn/rest/util/NoOpRecordingShutdownHandler.java
@@ -18,7 +18,7 @@
  */
 package org.apache.brooklyn.rest.util;
 
-import org.apache.brooklyn.rest.util.ShutdownHandler;
+import org.apache.brooklyn.core.mgmt.ShutdownHandler;
 
 public class NoOpRecordingShutdownHandler implements ShutdownHandler {
     private volatile boolean isRequested;
